requirements.txt
requests
wget
argparse
colorama
humanfriendly
phonenumbers
scapy
discord.py==1.6.0

